h2 impl

+++ backend/src/main/java/com/aleleone/WOD/Randomizer/presentation/controller/ExerciseController.java
@@ -29,7 +29,7 @@ public class ExerciseController {
 	}
 	
 	@GetMapping("/users/{username}/exercises/{id}")
-	public Exercise getExercise(@PathVariable String username, @PathVariable int id) {
+	public Exercise getExercise(@PathVariable String username, @PathVariable Long id) {
 		return exerciseService.getExerciseForUsername(username, id);
 	}
 	
@@ -47,7 +47,7 @@ public class ExerciseController {
 	
 	
 	@PutMapping("/users/{username}/exercises/{id}")
-	public Exercise updateExercise(@PathVariable String username, @PathVariable int id, @RequestBody Exercise exercise) throws UsernameNotFoundException {
+	public Exercise updateExercise(@PathVariable String username, @PathVariable Long id, @RequestBody Exercise exercise) throws UsernameNotFoundException {
 		return exerciseService.updateExercise(username, id, exercise);
 	}
 }
